@@ -1,7 +1,7 @@
 package com.algaworks.algafood.api.v1.model.input;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
 import org.springframework.http.MediaType;
 import org.springframework.web.multipart.MultipartFile;
