@@ -1,7 +1,7 @@
 package com.algaworks.algafood.api.v1.model.input;
 
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.PositiveOrZero;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.PositiveOrZero;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
