@@ -2,9 +2,9 @@ package com.algaworks.algafood.api.v1.model.input;
 
 import java.math.BigDecimal;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.PositiveOrZero;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.PositiveOrZero;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
