@@ -2,7 +2,7 @@ package com.algaworks.algafood.infrastructure.repository.spec;
 
 import java.util.ArrayList;
 
-import javax.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Predicate;
 
 import org.springframework.data.jpa.domain.Specification;
 
@@ -21,11 +21,11 @@ public class PedidoSpecs {
 			var predicates = new ArrayList<Predicate>();
 			
 			if (filtro.getClienteId() != null) {
-				predicates.add(builder.equal(root.get("cliente"), filtro.getClienteId()));
+				predicates.add(builder.equal(root.get("cliente").get("id"), filtro.getClienteId()));
 			}
 			
 			if (filtro.getRestauranteId() != null) {
-				predicates.add(builder.equal(root.get("restaurante"), filtro.getRestauranteId()));
+				predicates.add(builder.equal(root.get("restaurante").get("id"), filtro.getRestauranteId()));
 			}
 			
 			if (filtro.getDataCriacaoInicio() != null) {
