@@ -1,7 +1,7 @@
 package com.algaworks.algafood.infrastructure.service.email;
 
-import javax.mail.MessagingException;
-import javax.mail.internet.MimeMessage; 
+import jakarta.mail.MessagingException;
+import jakarta.mail.internet.MimeMessage; 
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.javamail.MimeMessageHelper;
