@@ -4,11 +4,11 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Root;
 
 import org.springframework.stereotype.Repository;
 
@@ -56,7 +56,7 @@ public class VendaQueryServiceImpl implements VendaQueryService {
 		var predicates = new ArrayList<Predicate>();
 		
 		if (filtro.getRestauranteId() != null) {
-		    predicates.add(builder.equal(root.get("restaurante"), filtro.getRestauranteId()));
+		    predicates.add(builder.equal(root.get("restaurante").get("id"), filtro.getRestauranteId()));
 		}
 		    
 		if (filtro.getDataCriacaoInicio() != null) {
