@@ -11,6 +11,7 @@ import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.security.access.AccessDeniedException;
@@ -49,17 +50,9 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	
 	@Override
 	protected ResponseEntity<Object> handleHttpMediaTypeNotAcceptable(HttpMediaTypeNotAcceptableException ex,
-			HttpHeaders headers, HttpStatus status, WebRequest request) {
+			HttpHeaders headers, HttpStatusCode status, WebRequest request) {
 		return ResponseEntity.status(status).headers(headers).build();
-	}
-	
-	@Override
-	protected ResponseEntity<Object> handleBindException(BindException ex, HttpHeaders headers, HttpStatus status,
-			WebRequest request) {
-		
-		return handleValidationInternal(ex, ex.getBindingResult(), headers, status, request);
-	}
-	
+	}	
 	
 	@ExceptionHandler({ ValidacaoException.class })
 	public ResponseEntity<Object> handleValidacaoException(ValidacaoException ex, WebRequest request) {
@@ -69,7 +62,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	
 	@Override
 	protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex,
-	        HttpHeaders headers, HttpStatus status, WebRequest request) {
+	        HttpHeaders headers, HttpStatusCode status, WebRequest request) {
 	    return handleValidationInternal(ex, ex.getBindingResult(), headers, status, request);
 	}
 	
@@ -85,7 +78,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	// 5. Poderíamos fazer tudo dentro de handleTypeMismatch, mas preferi separar em outro método
 	@Override
 	protected ResponseEntity<Object> handleTypeMismatch(TypeMismatchException ex, HttpHeaders headers,
-	        HttpStatus status, WebRequest request) {
+			HttpStatusCode status, WebRequest request) {
 	    
 	    if (ex instanceof MethodArgumentTypeMismatchException) {
 	        return handleMethodArgumentTypeMismatch(
@@ -97,7 +90,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	
 	private ResponseEntity<Object> handleMethodArgumentTypeMismatch(
 	        MethodArgumentTypeMismatchException ex, HttpHeaders headers,
-	        HttpStatus status, WebRequest request) {
+	        HttpStatusCode status, WebRequest request) {
 
 	    ProblemType problemType = ProblemType.PARAMETRO_INVALIDO;
 
@@ -114,7 +107,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	
 	@Override
 	protected ResponseEntity<Object> handleNoHandlerFoundException(NoHandlerFoundException ex, 
-	        HttpHeaders headers, HttpStatus status, WebRequest request) {
+	        HttpHeaders headers, HttpStatusCode status, WebRequest request) {
 	    
 	    ProblemType problemType = ProblemType.RECURSO_NAO_ENCONTRADO;
 	    String detail = String.format("O recurso %s, que você tentou acessar, é inexistente.", 
@@ -129,7 +122,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	
 	@Override
 	protected ResponseEntity<Object> handleHttpMessageNotReadable(HttpMessageNotReadableException ex,
-			HttpHeaders headers, HttpStatus status, WebRequest request) {
+			HttpHeaders headers, HttpStatusCode status, WebRequest request) {
 		Throwable rootCause = ExceptionUtils.getRootCause(ex);
 		
 	    if (rootCause instanceof InvalidFormatException) {
@@ -149,7 +142,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	}
 		
 	private ResponseEntity<Object> handleInvalidFormat(InvalidFormatException ex,
-			HttpHeaders headers, HttpStatus status, WebRequest request) {
+			HttpHeaders headers, HttpStatusCode status, WebRequest request) {
 
 		String path = joinPath(ex.getPath());
 		
@@ -166,7 +159,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	}
 	
 	private ResponseEntity<Object> handlePropertyBinding(PropertyBindingException ex,
-	        HttpHeaders headers, HttpStatus status, WebRequest request) {
+	        HttpHeaders headers, HttpStatusCode status, WebRequest request) {
 
 	    // Criei o método joinPath para reaproveitar em todos os métodos que precisam
 	    // concatenar os nomes das propriedades (separando por ".")
@@ -265,13 +258,13 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	
 	@Override
 	protected ResponseEntity<Object> handleExceptionInternal(Exception ex, Object body, HttpHeaders headers,
-			HttpStatus status, WebRequest request) {
+			HttpStatusCode status, WebRequest request) {
 		
 		if (body == null) {
 			body = Problem.builder()
 				.timestamp(OffsetDateTime.now())
-				.title(status.getReasonPhrase())
-				.status(status.value())
+				.title(HttpStatus.valueOf(status.value()).getReasonPhrase())
+				.status(status.value()) 
 				.userMessage(MSG_ERRO_GENERICA_USUARIO_FINAL)
 				.build();
 		} else if (body instanceof String) {
@@ -286,7 +279,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 		return super.handleExceptionInternal(ex, body, headers, status, request);
 	}
 	
-	private Problem.ProblemBuilder createProblemBuilder(HttpStatus status,
+	private Problem.ProblemBuilder createProblemBuilder(HttpStatusCode status,
 			ProblemType problemType, String detail) {
 		
 		return Problem.builder()
@@ -304,7 +297,7 @@ public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 	}
 	
 	private ResponseEntity<Object> handleValidationInternal(Exception ex, BindingResult bindingResult, HttpHeaders headers,
-			HttpStatus status, WebRequest request) {
+			HttpStatusCode status, WebRequest request) {
 		        
 	    ProblemType problemType = ProblemType.DADOS_INVALIDOS;
 	    String detail = "Um ou mais campos estão inválidos. Faça o preenchimento correto e tente novamente.";
