@@ -25,7 +25,13 @@ public class AlgaSecurity {
 	public Long getUsuarioId() {
 		Jwt jwt = (Jwt) getAuthentication().getPrincipal();
 		
-		return jwt.getClaim("usuario_id");
+		Object usuarioId = jwt.getClaim("usuario_id");
+		
+		if (usuarioId == null) {
+			return null;
+		}
+		
+		return Long.valueOf(usuarioId.toString());
 	}
 	
 	public boolean gerenciaRestaurante(Long restauranteId) {
