@@ -3,11 +3,9 @@ package com.algaworks.algafood.domain.model.dto;
 import java.math.BigDecimal;
 import java.util.Date;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-@AllArgsConstructor
 @Setter
 @Getter
 public class VendaDiaria {
@@ -15,5 +13,12 @@ public class VendaDiaria {
 	private Date data;
 	private Long totalVendas;
 	private BigDecimal totalFaturado;
+	
+	
+	public VendaDiaria(java.sql.Date data, Long totalVendas, BigDecimal totalFaturado) {
+		this.data = new Date(data.getTime());
+		this.totalVendas = totalVendas;
+		this.totalFaturado = totalFaturado;
+	}
 	
 }
